requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
